requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pyodbc==5.0.1
pymssql==2.2.7
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
joblib==1.3.2
rapidfuzz==3.6.1
